requirements.txt
pip~=25.1.1
attrs~=25.3.0
distro~=1.9.0
PySocks~=1.7.1
protobuf~=6.31.1
filelock~=3.18.0
fastapi~=0.115.14
faster-whisper~=1.1.1
python-multipart~=0.0.9

pydantic~=2.11.7
openai~=1.97.0
python-dotenv~=1.1.1